# docs/swagger.yaml
definitions:
...
  json.HistoricoVeiculoResponse:
    properties:
      data:
        type: string
      data_cadastro:
        type: string
      descricao:
        type: string
      id:
        type: integer
      id_anuncio:
        type: integer
      id_veiculo:
        type: integer
    type: object
...
  json.HistoricosVeiculoResponse:
    properties:
      historicos:
        items:
          $ref: '#/definitions/json.HistoricoVeiculoResponse'
        type: array
      total:
        type: integer
    type: object
...
  json.VeiculoRequest:
    properties:
      ano:
        maximum: 2030
        minimum: 1900
        type: integer
      cor:
        type: string
      id_usuario:
        type: integer
      modelo:
        type: string
      placa:
        type: string
    required:
    - ano
    - cor
    - id_usuario
    - modelo
    - placa
    type: object
  json.VeiculoResponse:
    properties:
      ano:
        type: integer
      ativo:
        type: boolean
      cor:
        type: string
      data_cadastro:
        type: string
      id:
        type: integer
      id_usuario:
        type: integer
      modelo:
        type: string
      placa:
        type: string
    type: object
...
  json.VeiculosResponse:
    properties:
      total:
        type: integer
      veiculos:
        items:
          $ref: '#/definitions/json.VeiculoResponse'
        type: array
    type: object
info:
  contact: {}
paths:
...
  /usuarios/{id_usuario}/historico:
    get:
      consumes:
      - application/json
      description: Retorna o histórico de todos os veículos de um usuário específico
      parameters:
      - description: ID do usuário
        in: path
        name: id_usuario
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/json.HistoricosVeiculoResponse'
        "400":
          description: ID de usuário inválido
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Lista histórico de todos os veículos de um usuário
      tags:
      - Veículos
  /usuarios/{id_usuario}/veiculos:
    get:
      consumes:
      - application/json
      description: Retorna todos os veículos ativos de um usuário específico
      parameters:
      - description: ID do usuário
        in: path
        name: id_usuario
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/json.VeiculosResponse'
        "400":
          description: ID de usuário inválido
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Lista veículos de um usuário
      tags:
      - Veículos
...
  /veiculos:
    get:
      consumes:
      - application/json
      description: Retorna uma lista com todos os veículos ativos
      produces:
      - application/json
      responses:
        "200":
          description: Lista de veículos
          schema:
            items:
              $ref: '#/definitions/json.VeiculoResponse'
            type: array
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Lista todos os veículos
      tags:
      - Veículos
    post:
      consumes:
      - application/json
      description: Cria um novo veículo com todos os dados fornecidos
      parameters:
      - description: Dados completos do veículo
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/json.VeiculoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Veículo criado com sucesso
          schema:
            $ref: '#/definitions/json.VeiculoResponse'
        "400":
          description: Dados inválidos
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Criação do veículo completo
      tags:
      - Veículos
...
  /veiculos/{id}:
    delete:
      consumes:
      - application/json
      description: Realiza soft delete do veículo, marcando como excluído sem remover
        do banco
      parameters:
      - description: ID do veículo
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Veículo excluído com sucesso
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID inválido
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Veículo não encontrado
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Exclui veículo (soft delete)
      tags:
      - Veículos
    get:
      consumes:
      - application/json
      description: Retorna os dados de um veículo específico pelo ID
      parameters:
      - description: ID do veículo
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Veículo encontrado
          schema:
            $ref: '#/definitions/json.VeiculoResponse'
        "400":
          description: ID inválido
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Veículo não encontrado
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Busca veículo por ID
      tags:
      - Veículos
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um veículo existente
      parameters:
      - description: ID do veículo
        in: path
        name: id
        required: true
        type: integer
      - description: Dados atualizados do veículo
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/json.VeiculoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Veículo atualizado com sucesso
          schema:
            $ref: '#/definitions/json.VeiculoResponse'
        "400":
          description: Dados inválidos
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Veículo não encontrado
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Atualiza veículo
      tags:
      - Veículos
...
  /veiculos/{id}/historico:
    get:
      consumes:
      - application/json
      description: Retorna o histórico completo de um veículo específico
      parameters:
      - description: ID do veículo
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/json.HistoricosVeiculoResponse'
        "400":
          description: ID de veículo inválido
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Lista histórico de um veículo
      tags:
      - Veículos
  /veiculos/{id}/restore:
    post:
      consumes:
      - application/json
      description: Restaura um veículo que foi soft deleted
      parameters:
      - description: ID do veículo
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Veículo restaurado com sucesso
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID inválido
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Veículo não encontrado
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Restaura veículo excluído
      tags:
      - Veículos
swagger: "2.0"